fix: replace servers
